WIP: child-rekey: Remove redundant migrate() call for child-create sub-task
When retrying due to a DH group mismatch this is already done by the
child-create task itself.  And in other cases where the task returns
NEED_MORE we actually will need access to a possible CHILD_SA to properly
delete it.

WIP: Error handling does not actually work quite right. For many
errors we get an INFORMATIONAL exchange with a REKEY_SA notify.