auth-cfg: Similar to certificates matching one CA should be enough
Not sure if defining multiple CA constraints and enforcing _all_ of them,
that is, the previous behavior, makes even sense.  To ensure a very specific
chain it should be enough to define the last intermediate CA.  On the
other hand, the ability to define multiple CAs could simplify configuration.

This can currently only be used with swanctl/VICI based configs as `rightca`
only takes a single DN.